made login available by email or username, moved businessSchema to its own model and updated exports
