+++ server/models/index.js
@@ -1,4 +1,5 @@
 const User = require('./Profile');
 const Review = require('./Review');
+const Business = require('./Business');
 
-module.exports = { User, Review };
+module.exports = { User, Review, Business };
